+++ src/main/scala/de/htwg/se/twothousandfortyeight/controller/actorDslImpl/Move.scala
@@ -1,4 +1,4 @@
-package de.htwg.se.twothousandfortyeight.controller.actorBaseImpl
+package de.htwg.se.twothousandfortyeight.controller.actorDslImpl
 
 case class Move(seq: String*) {
   seq.foreach(e => Player.play(e))
