+++ src/main/scala/de/htwg/se/twothousandfortyeight/examples/Game.sc
@@ -1,3 +1,3 @@
-import de.htwg.se.twothousandfortyeight.controller.actorBaseImpl.Move
+import de.htwg.se.twothousandfortyeight.controller.actorDslImpl.Move
 
 Move("w", "a", "s", "d")
